Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/httpcomponents/httpclient/trunk@939854 13f79535-47bb-0310-9956-ffa450edef68

+++ httpclient-cache/src/main/java/org/apache/http/client/cache/impl/CachedHttpResponseGenerator.java
@@ -60,7 +60,7 @@ public class CachedHttpResponseGenerator {
 
         long age = entry.getCurrentAgeSecs();
         if (age > 0) {
-            if (age >= (long) Integer.MAX_VALUE) {
+            if (age >= Integer.MAX_VALUE) {
                 response.setHeader(HeaderConstants.AGE, "2147483648");
             } else {
                 response.setHeader(HeaderConstants.AGE, "" + ((int) age));
